Prevent creation of project with spaces in their name
Gerrit allows spaces in project names but this is not well supported.
Some part of the code is not working with spaces, e.g. reindex-if-stale
task is always failing.

Even if would support project with spaces perfectly, we still want to
forbid creating project with spaces for various reasons:

-Confusion: project starting with a space will not be kept by Git in the
cloned folder name.
-On server side git folder, admins need to deal with spaces, especially
in any scripts.
-...
Change-Id: Ia55617781fc1cd6a1a0a6b35533b75b4a264c4f5