Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+2
> 
> Optional: Squash this and the previous change into the 2.16 update?

I'd prefer not to, as I wanted the latter to be as clear as possible, for the review. Meaning, showing only the 2.16-porting diffs, without consequent refactorings such as this change and its parent.

I also wanted to split this specific refactoring from the parent one, again for focused commit/review purposes. However I'm squashing Iaf7c1b8f now as it is a byproduct of the 2.16 update verification in Eclipse.

Patch-set: 3