Introduce an optional configuration to define a regex for naming rules
Before this change, this plugins enforced a rule that a project cannot
contain space in its name. Therefore, projects can have a name with any
set of characters while they don't include a space in their names.
However, due to IRI (not URI) and backward compatibility, the non-ASCII
characters may be represented in a way that can cause issues [1].

This change introduces a configurable parameter that can be set to
define a regex that can be enforced upon project creation to ensure
backward compatibility. If not defined, the default behavior, to check
that projects doesn't contain space, applies.

If the regex is defined, it will apply if it accepts the character /,
else it will get neglected to not affect the functionallity of the
plugin.

[1] https://en.wikipedia.org/wiki/Internationalized_Resource_Identifier

Bug: Issue 14297
Change-Id: Ia21c9834f41c4d5289f56aacd357c5de09f4521f